Fixed several bugs

+++ packages/fsdrv/ufs.lua
@@ -114,6 +114,7 @@ end
 ufs.makeDir = function(mountPath, device, path)
   path = fsd.resolveLinks(path)
   path = fsd.stripPath(mountPath, path)
+  if fsd.normalizePath(path) == "/UFSDATA" then error("Internal error") return end
   oldfs.makeDir(device .. path)
   fs.setNode(mountPath .. "/" .. path)
 end
@@ -123,6 +124,8 @@ ufs.move = function(mountPath, device, from, to)
   to = fsd.resolveLinks(to)
   from = fsd.stripPath(mountPath, from)
   to = fsd.stripPath(mountPath, to)
+  if fsd.normalizePath(to) == "/UFSDATA" then error("Internal error") return end
+  if fsd.normalizePath(from) == "/UFSDATA" then error("Internal error") return end
   oldfs.move(device .. from, device .. to)
   fs.setNode(mountPath .. "/" .. to)
 end
@@ -132,6 +135,8 @@ ufs.copy = function(mountPath, device, from, to)
   to = fsd.resolveLinks(to)
   from = fsd.stripPath(mountPath, from)
   to = fsd.stripPath(mountPath, to)
+  if fsd.normalizePath(to) == "/UFSDATA" then error("Internal error") return end
+  if fsd.normalizePath(from) == "/UFSDATA" then error("Internal error") return end
   oldfs.copy(device .. from, device .. to)
   fs.setNode(mountPath .. "/" .. to)
 end
@@ -139,6 +144,7 @@ end
 
 ufs.delete = function(mountPath, device, path)
   path = fsd.stripPath(mountPath, path)
+  if fsd.normalizePath(path) == "/UFSDATA" then error("Internal error") return end
   fsd.setNode(path, nil, nil, false)
   oldfs.delete(device .. path)
   fs.deleteNode(mountPath .. "/" .. path)
